refactor(pkg/logger): move to embedded folder
- move logger from pkg to embedded
- fix logger pkg import names
