+++ cmd/immuclient/audit/auditagent.go
@@ -26,10 +26,10 @@ import (
 	immusrvc "github.com/codenotary/immudb/cmd/sservice"
 	"github.com/codenotary/immudb/pkg/server"
 
+	"github.com/codenotary/immudb/embedded/logger"
 	"github.com/codenotary/immudb/pkg/client"
 	"github.com/codenotary/immudb/pkg/client/auditor"
 	"github.com/codenotary/immudb/pkg/client/state"
-	"github.com/codenotary/immudb/pkg/logger"
 	"github.com/spf13/viper"
 	"github.com/takama/daemon"
 )
