+++ embedded/logger/memory.go
