+++ embedded/logger/simple.go
