+++ embedded/tbtree/options.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/codenotary/immudb/embedded/appendable"
 	"github.com/codenotary/immudb/embedded/appendable/multiapp"
-	"github.com/codenotary/immudb/pkg/logger"
+	"github.com/codenotary/immudb/embedded/logger"
 )
 
 const DefaultMaxNodeSize = 4096
