+++ pkg/client/auditor/monitoring_server.go
@@ -22,8 +22,8 @@ import (
 	"expvar"
 	"net/http"
 
+	"github.com/codenotary/immudb/embedded/logger"
 	"github.com/codenotary/immudb/pkg/api/schema"
-	"github.com/codenotary/immudb/pkg/logger"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
