+++ pkg/client/heartbeater/heartbeater.go
@@ -21,8 +21,8 @@ import (
 	stdos "os"
 	"time"
 
+	"github.com/codenotary/immudb/embedded/logger"
 	"github.com/codenotary/immudb/pkg/api/schema"
-	"github.com/codenotary/immudb/pkg/logger"
 	"github.com/golang/protobuf/ptypes/empty"
 )
