+++ pkg/client/types.go
@@ -18,12 +18,13 @@ package client
 
 import (
 	"crypto/ecdsa"
+
 	"github.com/codenotary/immudb/pkg/client/tokenservice"
 	"github.com/codenotary/immudb/pkg/stream"
 
+	"github.com/codenotary/immudb/embedded/logger"
 	"github.com/codenotary/immudb/pkg/api/schema"
 	"github.com/codenotary/immudb/pkg/client/state"
-	"github.com/codenotary/immudb/pkg/logger"
 	"google.golang.org/grpc"
 )
