+++ pkg/pgsql/server/session_factory.go
@@ -18,9 +18,10 @@ package server
 
 import (
 	"crypto/tls"
-	"github.com/codenotary/immudb/pkg/database"
-	"github.com/codenotary/immudb/pkg/logger"
 	"net"
+
+	"github.com/codenotary/immudb/embedded/logger"
+	"github.com/codenotary/immudb/pkg/database"
 )
 
 type sessionFactory struct{}
