+++ pkg/server/server.go
@@ -42,7 +42,7 @@ import (
 
 	"github.com/codenotary/immudb/pkg/database"
 
-	"github.com/codenotary/immudb/pkg/logger"
+	"github.com/codenotary/immudb/embedded/logger"
 	"github.com/codenotary/immudb/pkg/signer"
 
 	"github.com/codenotary/immudb/cmd/helper"
