+++ webconsole/webconsole_default_test.go
@@ -1,16 +1,18 @@
+//go:build !webconsole
 // +build !webconsole
 
 package webconsole
 
 import (
-	"github.com/codenotary/immudb/pkg/logger"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"testing"
+
+	"github.com/codenotary/immudb/embedded/logger"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestSetupWebconsoleDefault(t *testing.T) {
